:truck: api: move getPrivateCertificate method from lib to src
