+++ api/src/certification/results/infrastructure/repositories/certificate-repository.js
@@ -79,10 +79,35 @@ const findPrivateCertificatesByUserId = async function ({ userId }) {
   return privateCertificates;
 };
 
+const getPrivateCertificate = async function (id, { locale } = {}) {
+  const certificationCourseDTO = await _selectPrivateCertificates()
+    .where('certification-courses.id', '=', id)
+    .groupBy('certification-courses.id', 'sessions.id', 'assessment-results.id')
+    .where('certification-courses.isPublished', true)
+    .where('certification-courses.isCancelled', false)
+    .where('assessment-results.status', AssessmentResult.status.VALIDATED)
+    .first();
+
+  if (!certificationCourseDTO) {
+    throw new NotFoundError(`Certificate not found for ID ${id}`);
+  }
+
+  const certifiedBadges = await _getCertifiedBadges(id);
+
+  const competenceTree = await competenceTreeRepository.get({ locale });
+
+  return _toDomainForPrivateCertificate({
+    certificationCourseDTO,
+    competenceTree,
+    certifiedBadges,
+  });
+};
+
 export {
   findByDivisionForScoIsManagingStudentsOrganization,
   findPrivateCertificatesByUserId,
   getCertificationAttestation,
+  getPrivateCertificate,
 };
 
 function _selectCertificationAttestations() {
